Add response DTOs for Room and BookedRoom

+++ hotel-booking-backend/src/main/java/com/krlaan/hotelbooking/model/BookedRoom.java
@@ -17,7 +17,7 @@ public class BookedRoom {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int bookingId;
+    private Long bookingId;
 
     @Column(name = "check_in")
     private LocalDate checkInDate;
